extract infoButton module
